// src/utils/ct-types.d.ts
export type CustomModule = CustomModuleCreate & {
    id: number;
};

export type CustomModuleCreate = {
    description?: string;
    name: string;
    shorty: string;
    sortKey: number;
};

export type CustomModuleDataCategory = CustomModuleDataCategoryCreate & {
    id: number;
};

export type CustomModuleDataCategoryCreate = {
    customModuleId: number;
    data?: string;
    description: string;
    name: string;
    shorty: string;
};

export type CustomModuleDataValue = CustomModuleDataValueCreate & {
    id: number;
};

export type CustomModuleDataValueCreate = {
    dataCategoryId: number;
    value: string;
};